requirements.txt
h5py==3.10.0
imageio==2.27.0
lpips==0.1.4
matplotlib==3.7.1
ninja==1.11.1
numba==0.57.0
numpy==1.24.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pyiqa==0.1.6.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
tensorflow-gpu==2.9.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0